Fix a few more media element test
